// CSKT/originaldb-gate-fe/src/modules/techinical-organization/enum/LevelCategory.enum.ts
export enum LevelCategoriesScope {
  CHIEN_DICH = "CHIEN_DICH",
  CHIEN_LUOC = "CHIEN_LUOC",
  CHIEN_THUAT = "CHIEN_THUAT",
}
export const LevelCategoriesLabel: any = {
  CHIEN_DICH: "Chiến dịch",
  CHIEN_LUOC: "Chiến lược",
  CHIEN_THUAT: "Chiến thuật",
};
export const OfficalLabel: any = {
  KIEM_NHIEM: "Kiêm nhiệm",
  CHINH_THUC: "Chính thức",
};
export enum ModalType {
  CSBD = "CSBD",
  ToBD = "ToBD",
  TrCQDT = "TRAM_CQDT",
}
export enum TechinicalType {
  ToBDKT = "to_bdktcd",
  TrCQDT = "TRAM_CQDT",
  Xuong = "xuong",
  Kho = "kho",
  Tram = "tram",
  Cssc = "cssc_csdt",
  Cosdl = "csdl",
}